Add support for building placement, needs reload atm.

+++ apps/game/src/components/Collider.ts
@@ -3,8 +3,8 @@ import { Types, defineComponent } from "@virtcon2/virt-bit-ecs";
 export const Collider = defineComponent({
   sizeWidth: Types.i32,
   sizeHeight: Types.i32,
-  offsetX: Types.i32,
-  offsetY: Types.i32,
+  offsetX: Types.f32,
+  offsetY: Types.f32,
   scale: Types.f32,
   static: Types.ui8, // 0 = false, 1 = true
 });
